Extract SBNProvider to its own file
