+++ spec/alexandria/book_providers/sbn_provider_spec.rb
@@ -6,6 +6,8 @@
 
 require "spec_helper"
 
-describe Alexandria::BookProviders do
-  it "should be less clever"
+describe Alexandria::BookProviders::SBNProvider do
+  it "works" do
+    assert_correct_search_result(described_class, "9788835926436")
+  end
 end
